chore(GTM): move sdk loader to different file
