+++ src/integrations/GoogleTagManager/browser.js
@@ -1,7 +1,7 @@
 /* eslint-disable class-methods-use-this */
 import logger from '../../utils/logUtil';
-import { LOAD_ORIGIN } from '../../utils/ScriptLoader';
 import { NAME } from './constants';
+import { loader } from './loader';
 
 class GoogleTagManager {
   constructor(config, analytics, destinationInfo) {
@@ -26,17 +26,7 @@ class GoogleTagManager {
 
     window.finalUrl = this.serverUrl ? this.serverUrl : defaultUrl;
 
-    (function (w, d, s, l, i) {
-      w[l] = w[l] || [];
-      w[l].push({ 'gtm.start': new Date().getTime(), event: 'gtm.js' });
-      const f = d.getElementsByTagName(s)[0];
-      const j = d.createElement(s);
-      const dl = l !== 'dataLayer' ? `&l=${l}` : '';
-      j.setAttribute('data-loader', LOAD_ORIGIN);
-      j.async = true;
-      j.src = `${window.finalUrl}/gtm.js?id=${i}${dl}`;
-      f.parentNode.insertBefore(j, f);
-    })(window, document, 'script', 'dataLayer', this.containerID);
+    loader(window, document, 'script', 'dataLayer', this.containerID);
   }
 
   identify(rudderElement) {
